requirements.txt
selenium
pytest
allure-pytest
seletools
requests
faker